requirements.txt
pandas==1.1.3
seaborn==0.11.0
matplotlib==3.3.2
numpy==1.19.2
streamlit==0.79.0
torchvision==0.8.2
torch==1.7.1
Pillow==8.1.2
